% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_default_aes.R
\name{add_default_aes}
\alias{add_default_aes}
\title{Add default aestethic settings to the variables and flows data frames.
Helper function called by prepare_diagram}
\usage{
add_default_aes(variables, flows)
}
\arguments{
\item{variables}{The variables data frame.}

\item{flows}{The flows data frame.}
}
\value{
A list of the two data frames.
}
\description{
Add default aestethic settings to the variables and flows data frames.
Helper function called by prepare_diagram
}
