% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plus_signs.R
\name{add_plus_signs}
\alias{add_plus_signs}
\title{Add explicit plus signs to flows. Used by prepare_diagram.}
\usage{
add_plus_signs(flows)
}
\arguments{
\item{flows}{A list of flows. Each list element is a vector of flows for a specific variable.}
}
\value{
The same list of flows that was sent into the function, but with explicit plus signs in front of positive flow terms.
}
\description{
Add explicit plus signs to flows. Used by prepare_diagram.
}
