% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_default_aesthetics.R
\name{apply_default_aesthetics}
\alias{apply_default_aesthetics}
\title{Applies default aesthetic values to data frames}
\usage{
apply_default_aesthetics(diagram_list)
}
\arguments{
\item{diagram_list}{A list of data frames.}
}
\value{
A list of data frames.
}
\description{
\code{apply_default_aesthetics} applies the default ggplot2 aesthetics from
\code{make_diagram()}. This is an internal helper function.
}
