% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataframes.R
\name{check_dataframes}
\alias{check_dataframes}
\title{Helper function to perform basic checks of the variables
and flows data frames. Returns errors if found.}
\usage{
check_dataframes(diagram_list)
}
\arguments{
\item{diagram_list}{The list of \code{variables} and \code{flows} data frames.}
}
\value{
Either \code{NULL} or an error message.
}
\description{
Helper function to perform basic checks of the variables
and flows data frames. Returns errors if found.
}
