% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_list.R
\name{check_model_list}
\alias{check_model_list}
\title{Check model_list input for prepare_diagram for correctness.}
\usage{
check_model_list(model_list)
}
\arguments{
\item{model_list}{the model input structure needed for the \code{\link{prepare_diagram}} function}
}
\value{
Either an error or nothing.
}
\description{
This function makes sure that the model_list structure
contains a properly specified model.
This is an internal function.
}
