% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_settings.R
\name{check_model_settings}
\alias{check_model_settings}
\title{Check optional model_settings input for prepare_diagram for correctness.}
\usage{
check_model_settings(model_list, model_settings)
}
\arguments{
\item{model_list}{model list structure, required input}

\item{model_settings}{additional settings, optional input}
}
\value{
Either an error message or null
}
\description{
This function checks the model_settings input argument for \code{\link{prepare_diagram}}.
This is mainly an internal function, called by \code{\link{prepare_diagram}}.
}
