% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_down_rows.R
\name{fill_down_rows}
\alias{fill_down_rows}
\title{Fill in NA rows with the first element.}
\usage{
fill_down_rows(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame.
}
\description{
This is a helper function that fills in empty rows
of a dataframe. The key assumption is that either all rows will be
filled or only the first row will be filled. Therefore, all NA
rows are given the value of the first element.
}
