% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_arrow_pos.R
\name{fix_arrow_pos}
\alias{fix_arrow_pos}
\title{Adjust arrow positions to avoid overlaps.
helper function for prepare_diagram that is only applied to
arrows from nowhere or to nowhere (e.g., birth or death arrows).}
\usage{
fix_arrow_pos(edf)
}
\arguments{
\item{edf}{The edges (flows) data frame.}
}
\description{
Adjust arrow positions to avoid overlaps.
helper function for prepare_diagram that is only applied to
arrows from nowhere or to nowhere (e.g., birth or death arrows).
}
