% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowdiagramr.R
\docType{package}
\name{flowdiagramr}
\alias{flowdiagramr}
\title{flowdiagramr: A package to create flow diagrams}
\description{
The flowdiagramr package allows users to quickly produce ggplot2
based model diagrams
}
