% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diagram.R
\name{make_diagram}
\alias{make_diagram}
\title{Make a ggplot2 model diagram.}
\usage{
make_diagram(diagram_list, with_grid = FALSE)
}
\arguments{
\item{diagram_list}{A required list of data frames returned from the
\code{prepare_diagram} function. See that function for details
about this object.}

\item{with_grid}{A logical indicating whether to return the ggplot
with a grid. Default is FALSE. The grid can be helpful if you
want/need to move items around.}
}
\value{
A ggplot2 object.
}
\description{
\code{make_diagram()} generates a \strong{ggplot2} object based on the data frames
made with \code{\link{prepare_diagram}}. The function only applies
aesthetics that are not associated with x, y locations. Colors, linetypes,
and other graphical options can be set by the user.
}
\examples{
mymodel = list(variables = c("S","I","R"),
               flows = list(S_flows = c("-b*S*I"),
                            I_flows = c("b*S*I","-g*I"),
                            R_flows = c("g*I") ) )
diagram_list <- prepare_diagram(model_list = mymodel)

# make diagram without grid
diagram <- make_diagram(diagram_list)

# make diagram with grid
diagram_with_grid <- make_diagram(diagram_list, with_grid = TRUE)

}
