% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diagram_inputs.R
\name{make_diagram_inputs}
\alias{make_diagram_inputs}
\title{Convert modelbuilder list object to list for modeldiagram}
\usage{
make_diagram_inputs(mbmodel)
}
\value{
A list.
}
\description{
Convert modelbuilder list object to list for modeldiagram
}
