% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_settings_df.R
\name{make_new_settings_df}
\alias{make_new_settings_df}
\title{Updates specified columns in a data frame with new values from a list.}
\usage{
make_new_settings_df(n, diagram_settings)
}
\arguments{
\item{n}{The number of variables or flows.}

\item{diagram_settings}{The diagram_settings list.}
}
\description{
Helper function that updates all settings columns in a
flows or variables data frame based on the new values in diagram_settings.
}
