% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vdf_angled.R
\name{make_vdf_angled}
\alias{make_vdf_angled}
\title{Helper function to adjust vertical edge start and end positions
to create angles.
helper function for prepare_diagram}
\usage{
make_vdf_angled(edf, ndf, model_settings)
}
\arguments{
\item{edf}{The edges (flows) data frame.}

\item{ndf}{The nodes (variables) data frame.}

\item{model_settings}{The model_settings list.}
}
\value{
A dataframe.
}
\description{
Helper function to adjust vertical edge start and end positions
to create angles.
helper function for prepare_diagram
}
