% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycle_values.R
\name{recycle_values}
\alias{recycle_values}
\title{Recycle values to certain length}
\usage{
recycle_values(x, n)
}
\arguments{
\item{x}{The values to potentially recycle.}

\item{n}{The desired length (i.e., number of items).}
}
\description{
Recycle values to certain length
}
