% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na_rows.R
\name{remove_na_rows}
\alias{remove_na_rows}
\title{Remove rows with no location information.}
\usage{
remove_na_rows(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame.
}
\description{
Remove rows with no location information.
}
