% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_curvature.R
\name{set_curvature}
\alias{set_curvature}
\title{Sets the curvature values for curved arrows. used by prepare_diagram}
\usage{
set_curvature(edf, ndf)
}
\arguments{
\item{edf}{Edges df}

\item{ndf}{Nodes df}
}
\value{
A data frame
export
}
\description{
Sets the curvature values for curved arrows. used by prepare_diagram
}
