% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_node_to_na.R
\name{set_node_to_na}
\alias{set_node_to_na}
\title{Set to and from values to NA if value not present in nodes df}
\usage{
set_node_to_na(targdf, nodedf)
}
\arguments{
\item{targdf}{A data frame.}

\item{nodedf}{A data frame.}
}
\description{
Set to and from values to NA if value not present in nodes df
}
