% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_setting_lengths.R
\name{test_setting_lengths}
\alias{test_setting_lengths}
\title{Test whether settings are the correct length.}
\usage{
test_setting_lengths(diagram_settings, names, n)
}
\arguments{
\item{diagram_settings}{The diagram_settings list.}

\item{names}{The relevant list names for the test.}

\item{n}{The number of values allowed, other than 1.}
}
\value{
A logical.
}
\description{
Helper function called within update_settings to test
whether the number of values provided for each setting are either
of length 1 or the length of the flows or variables.
}
