% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_diagram.R
\name{update_diagram}
\alias{update_diagram}
\title{Updates the aesthetics settings in the variables and flows data frames}
\usage{
update_diagram(diagram_list, diagram_settings = NULL)
}
\arguments{
\item{diagram_list}{A required list of data frames returned from the
\code{\link{prepare_diagram}} function. See that function for details
about this object.}

\item{diagram_settings}{A required list of diagram aesthetic settings. The
following elements are supported and default values are provided:
\itemize{
\item \code{var_xmin}: A numeric vector of minimum x locations for each variable
in the \code{variables} data frame. The vector must equal the number of
variables and is assumed to be in the same order as the variables in the
\code{variables} data frame.
\item \code{var_xmax}: A numeric vector of maximum x locations for each variable
in the \code{variables} data frame. The vector must equal the number of
variables and is assumed to be in the same order as the variables in the
\code{variables} data frame.
\item \code{var_ymin}: A numeric vector of minimum y locations for each variable
in the \code{variables} data frame. The vector must equal the number of
variables and is assumed to be in the same order as the variables in the
\code{variables} data frame.
\item \code{var_ymax}: A numeric vector of maximum y locations for each variable
in the \code{variables} data frame. The vector must equal the number of
variables and is assumed to be in the same order as the variables in the
\code{variables} data frame.
\item \code{var_xlabel}: A numeric vector of the x location for the label for
each variable in the \code{variables} data frame. The vector must equal the
number of variables and is assumed to be in the same order as the
variables in the \code{variables} data frame.
\item \code{var_ylabel}: A numeric vector of the y location for the label for
each variable in the \code{variables} data frame. The vector must equal the
number of variables and is assumed to be in the same order as the
variables in the \code{variables} data frame.
\item \code{var_outline_color}: A character string or vector of character strings
specifying the color of variable outlines. Must either be of length 1
or the number of rows in the variables data frame.
\item \code{var_fill_color}: A character string or vector of character strings
specifying the fill color of variables. Must either be of length 1
or the number of rows in the variables data frame.
\item \code{var_label_color}: A character string or vector of character strings
specifying the text color for variable labels. Must either be of length 1
or the number of rows in the variables data frame.
\item \code{var_label_size}: A numeric scalar specifying the text size for variable
labels. Must either be of length 1 or the number of rows in the
variables data frame.

\item \code{flow_xmin}: A numeric vector of minimum x locations for each flow
in the \code{flows} data frame. The vector must equal the number of
flows and is assumed to be in the same order as the flows in the
\code{flows} data frame. xmin is the starting point of the flow.
\item \code{flow_xmax}: A numeric vector of maximum x locations for each flow
in the \code{flows} data frame. The vector must equal the number of
flows and is assumed to be in the same order as the flows in the
\code{flows} data frame. xmax is the ending point of the flow.
\item \code{flow_ymin}: A numeric vector of minimum y locations for each flow
in the \code{flows} data frame. The vector must equal the number of
flows and is assumed to be in the same order as the flows in the
\code{flows} data frame. xmin is the starting point of the flow.
\item \code{flow_ymax}: A numeric vector of maximum y locations for each flow
in the \code{flows} data frame. The vector must equal the number of
flows and is assumed to be in the same order as the flows in the
\code{flows} data frame. xmax is the ending point of the flow.
\item \code{flow_xlabel}: A numeric vector of the x location for the label for
each flow in the \code{flows} data frame. The vector must equal the
number of flows and is assumed to be in the same order as the
flows in the \code{flows} data frame.
\item \code{flow_ylabel}: A numeric vector of the y location for the label for
each flow in the \code{flows} data frame. The vector must equal the
number of flows and is assumed to be in the same order as the
flows in the \code{flows} data frame.
\item \code{flow_curvature}: A numeric vector of curvature values for each flow
in the \code{flows} data frame. The vector must equal the
number of flows and is assumed to be in the same order as the
flows in the \code{flows} data frame.

\item \code{main_flow_line_color}: A character string or vector of character strings
specifying the text color for non-interaction flow arrows. Must either
be of length 1 or the number of main flows.
\item \code{main_flow_linetype}: Either a numeric scalar/vector or a character scalar/vector
specifying the linetype for main flows (non-interaction flows). This
argument is passed to the \code{linetype} argument in ggplot2. From
the ggplot2 documentation: "The linetype aesthetic can be specified
with either an integer (0-6), a name (0 = blank, 1 = solid, 2 = dashed,
3 = dotted, 4 = dotdash, 5 = longdash, 6 = twodash), a mapping to a
discrete variable, or a string of an even number (up to eight) of
hexadecimal digits which give the lengths in consecutive positions in
the string." Default is 1 (solid). Must either
be of length 1 or the number of main flows.
\item \code{main_flow_line_size}: A numeric scalar or vector specifying the line size for the
main flows (non-interaction flows). Must either
be of length 1 or the number of main flows.
\item \code{main_flow_label_color}: A character string or vector of character strings
specifying the text color for main flow labels. Must either
be of length 1 or the number of main flows.
\item \code{main_flow_label_size}: A scalar or numeric vector
specifying the text size for main flow labels. Must either
be of length 1 or the number of main flows.
\item \code{main_flow_show_label}: A logical scalar or vector indicating whether
the flow label should be shown (TRUE) or not (FALSE). Must either
be of length 1 or the number of main flows.
\item \code{main_flow_arrow_size}: A scalar or numeric vector
specifying the size for main flow arrow heads. Must either
be of length 1 or the number of main flows.
\item \code{main_flow_show_arrow}: A logical scalar or vector indicating whether
the entire flow arrow should be shown (TRUE) or not (FALSE). Must either
be of length 1 or the number of main flows.

\item \code{interaction_flow_line_color}: A character string or vector of
character strings specifying the text color for non-interaction flow
arrows. Must either be of length 1 or the number of interaction flows.
\item \code{interaction_flow_linetype}: Either a numeric scalar/vector or a
character scalar/vector specifying the linetype for interaction flows.
This argument is passed to the \code{linetype} argument in ggplot2. From
the ggplot2 documentation: "The linetype aesthetic can be specified
with either an integer (0-6), a name (0 = blank, 1 = solid, 2 = dashed,
3 = dotted, 4 = dotdash, 5 = longdash, 6 = twodash), a mapping to a
discrete variable, or a string of an even number (up to eight) of
hexadecimal digits which give the lengths in consecutive positions in
the string." Default is 1 (solid). Must either
be of length 1 or the number of interaction flows.
\item \code{interaction_flow_line_size}: A numeric scalar or vector specifying
the line size for the interaction flows (non-interaction flows). Must
either be of length 1 or the number of interaction flows.
\item \code{interaction_flow_label_color}: A character string or vector of
character strings specifying the text color for interaction flow labels.
Must either be of length 1 or the number of interaction flows.
\item \code{interaction_flow_label_size}: A scalar or numeric vector
specifying the text size for interaction flow labels. Must either
be of length 1 or the number of interaction flows.
\item \code{interaction_flow_show_label}: A logical scalar or vector indicating
whether the flow label should be shown (TRUE) or not (FALSE). Must
either be of length 1 or the number of interaction flows.
\item \code{interaction_flow_arrow_size}: A scalar or numeric vector
specifying the size for interaction flow arrow heads. Must either
be of length 1 or the number of interaction flows.
\item \code{interaction_flow_show_arrow}: A logical scalar or vector indicating whether
the entire flow arrow should be shown (TRUE) or not (FALSE). Must either
be of length 1 or the number of interaction flows.

\item \code{external_flow_line_color}: A character string or vector of character
strings specifying the text color for non-interaction flow arrows.
Must either be of length 1 or the number of external flows.
\item \code{external_flow_linetype}: Either a numeric scalar/vector or a character
scalar/vector specifying the linetype for external flows. This
argument is passed to the \code{linetype} argument in ggplot2. From
the ggplot2 documentation: "The linetype aesthetic can be specified
with either an integer (0-6), a name (0 = blank, 1 = solid, 2 = dashed,
3 = dotted, 4 = dotdash, 5 = longdash, 6 = twodash), a mapping to a
discrete variable, or a string of an even number (up to eight) of
hexadecimal digits which give the lengths in consecutive positions in
the string." Default is 1 (solid). Must either be of length 1 or
the number of external flows.
\item \code{external_flow_line_size}: A numeric scalar or vector specifying the
line size for the external flows (non-interaction flows). Must either be
of length 1 or the number of external flows.
\item \code{external_flow_label_color}: A character string or vector of character
strings specifying the text color for external flow labels. Must either
be of length 1 or the number of external flows.
\item \code{external_flow_label_size}: A scalar or numeric vector
specifying the text size for external flow labels. Must either be of
length 1 or the number of external flows.
\item \code{external_flow_show_label}: A logical scalar or vector indicating
whether the flow label should be shown (TRUE) or not (FALSE). Must
either be of length 1 or the number of external flows.
\item \code{external_flow_arrow_size}: A scalar or numeric vector
specifying the size for external flow arrow heads. Must either
be of length 1 or the number of external flows.
\item \code{external_flow_show_arrow}: A logical scalar or vector indicating whether
the entire flow arrow should be shown (TRUE) or not (FALSE). Must either
be of length 1 or the number of external flows.
}}
}
\value{
The same list of data frames that was sent into the function
with updated style settings (different values in the updated
columns/variables) of the data frames.
}
\description{
This function takes a list as created by \code{\link{prepare_diagram}}
and allows the user to update the styling of the boxes and arrows by setting
the different arguments in the input list of this function.
}
\details{
The user can provide any number of updates in the
\code{diagram_settings} list. Only those provided are updated. If the
function is called with no updates, a warning is issued and the original
data frames list object is returned.
}
\examples{
# basic model specification
variables <- c("S","I","R")
flows <- list(S_flows = c("-b*S*I"),
              I_flows = c("b*S*I","-g*I"),
              R_flows = c("g*I"))
mymodel <- list(variables = variables, flows = flows)
diag_list <- prepare_diagram(model_list = mymodel)
diag_list2 <- update_diagram(diag_list, diagram_settings = list(var_fill_color = "green"))



# more extensive updates
newsettings <- list(var_label_color = c("green","blue","red"),
                    interaction_flow_line_size = 1.5,
                    interaction_flow_line_color = "orange")
diag_list3 <- update_diagram(diag_list, diagram_settings = newsettings)

}
