% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_external_interaction_positions.R
\name{update_external_interaction_positions}
\alias{update_external_interaction_positions}
\title{Update the x,y positions for external interaction flows}
\usage{
update_external_interaction_positions(ext, ndf)
}
\arguments{
\item{ext}{The subsetted external interactions (extints).}

\item{ndf}{The variables (nodes) data frame.}
}
\value{
The extints data frame with location information.
}
\description{
Update the x,y positions for external interaction flows
}
