% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_interactions.R
\name{update_interactions}
\alias{update_interactions}
\title{Updates the interaction column for plotting.}
\usage{
update_interactions(df)
}
\arguments{
\item{df}{An edge data frame.}
}
\value{
The edge data frame.
}
\description{
\code{update_interactions()} sets the interaction column of an edge
data frame to TRUE if the \code{direct_interaction} columns is true. It
then removes the \code{direct_interaction} column because it is no longer
needed.
}
