% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_straight_labels.R
\name{update_straight_labels}
\alias{update_straight_labels}
\title{Updates the x position of labels running vertically.}
\usage{
update_straight_labels(flows)
}
\arguments{
\item{flows}{An edge data frame.}
}
\value{
The edge data frame.
}
\description{
Updates the x position of labels running vertically.
}
