% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tofroms.R
\name{update_tofroms}
\alias{update_tofroms}
\title{Updates the numeric to/from ids to the variable label}
\usage{
update_tofroms(flows, variables)
}
\arguments{
\item{flows}{The flows data frame.}

\item{variables}{The variables data frame.}
}
\description{
Updates the numeric to/from ids to the variable label
}
