% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_diagram.R
\name{write_diagram}
\alias{write_diagram}
\title{Write the ggplot2 code to an R file.}
\usage{
write_diagram(directory = NULL, filename = NULL)
}
\arguments{
\item{directory}{File directory in which to save the R file. Defualt
location is the current working directory.}

\item{filename}{Name of the file, must end in '.R'. Default name is
'diagram_ggplot_code.R'.}
}
\value{
A message telling the user where the file is.
}
\description{
Write the ggplot2 code to an R file.
}
