% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_diagram_code.R
\name{write_diagram_code}
\alias{write_diagram_code}
\title{Write the ggplot2 code.}
\usage{
write_diagram_code()
}
\value{
A ggplot2 code snippet as a character string.
}
\description{
Write the ggplot2 code.
}
