% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_locations.R
\name{add_locations}
\alias{add_locations}
\title{Add x,y location informations}
\usage{
add_locations(
  variables,
  varlocations = NULL,
  varbox_x_size,
  varbox_y_size,
  varspace_x_size,
  varspace_y_size
)
}
\arguments{
\item{variables}{The nodes (variables) data frame.}

\item{varlocations}{The varlocations matrix. Default is NULL.}

\item{varbox_x_size}{Scaler for box width, from \code{model_settings}.}

\item{varbox_y_size}{Scaler for box height, from \code{model_settings}.}

\item{varspace_x_size}{Scaler for horizontal spacing, from \code{model_settings}.}

\item{varspace_y_size}{Scaler for vertical spacing, from \code{model_settings}.}
}
\value{
The variables data frame with location information.
}
\description{
Add x,y location informations
}
