% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_from_modelbuilder.R
\name{convert_from_modelbuilder}
\alias{convert_from_modelbuilder}
\title{Convert modelbuilder list object to list for modeldiagram}
\usage{
convert_from_modelbuilder(mbmodel)
}
\arguments{
\item{mbmodel}{A \strong{modelbuilder} model object.}
}
\value{
A list object consisting of the two components used by
\code{\link{prepare_diagram}} function. The list contains two elements:
\itemize{
\item{\code{model_list}}: A list containing variable labels and flows.
This has the format needed for the first argument of \code{\link{prepare_diagram}}.
single uppercase characters. Must start with uppercase.
\item{\code{model_settings}}: A list containing the variable names.
Also sets \code{use_varnames = TRUE}. This has the format needed
for the second, optional argument of \code{\link{prepare_diagram}}.
}
}
\description{
\code{convert_from_modelbuilder()} takes a model, stored in an Rds file,
that was built with the \strong{modelbuilder} package, and
converts it to the format used by \strong{flowdiagramr}.
}
\examples{
\dontrun{
#loading a model file that was made with **modelbuilder**
mbmodel <- readRDS('SIR_model.Rds')
mbmodel_structure <- convert_from_modelbuilder(mbmodel)
}
}
