% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_from_modelbuilder.R
\name{convert_from_modelbuilder}
\alias{convert_from_modelbuilder}
\title{Convert modelbuilder list object to list for modeldiagram}
\usage{
convert_from_modelbuilder(mbmodel)
}
\arguments{
\item{mbmodel}{A modelbuilder object.}
}
\value{
A list object that can be used as input for the prepare_diagram function.
}
\description{
Convert modelbuilder list object to list for modeldiagram
}
