% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dataframes.R
\name{make_dataframes}
\alias{make_dataframes}
\title{Create data frames for plotting from model elements.}
\usage{
make_dataframes(input_list)
}
\arguments{
\item{input_list}{A list of model elements. Currently only accepts
a modelbuilder list object. At a minimum, the list must contain
two elements with names \code{vars} and \code{flows}.}
}
\value{
A list of data frames.
}
\description{
Create data frames for plotting from model elements.
}
