% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diagram.R
\name{make_diagram}
\alias{make_diagram}
\title{Make a ggplot2 model diagram.}
\usage{
make_diagram(
  diagram_list,
  diagram_settings = list(label_flows = TRUE, external_flows = TRUE, interaction_label
    = TRUE, var_outline_color = NA, var_fill_color = "#6aa4c8", var_text_color = "white",
    var_text_size = 10, flow_text_color = "black", flow_text_size = 5, main_arrow_color =
    "grey25", main_arrow_linetype = "solid", main_arrow_size = 0.7,
    interaction_arrow_color = "grey25", interaction_arrow_linetype = "dashed",
    interaction_arrow_size = 0.7, with_grid = FALSE)
)
}
\arguments{
\item{diagram_list}{A list of data frames returned from the
\code{prepare_diagram} function. See that function for details
about this object.}

\item{diagram_settings}{A list of diagram aesthetic settings. The
following elements are supported and default values are provided:
\itemize{
\item \code{label_flows}: A logical indicating whether to label the flows
(TRUE, default) or not (FALSE).
\item \code{external_flows}: A logical indicating whether to include flows into
and out of the system (external flows). Default is TRUE (include).
\item \code{interaction_label}: A logical indicating whether to make the diagram
with interaction terms (typically curved arrows leading to the
mid point of another arrow) or to simply label the main flow. See
vignettes for examples.
\item \code{node_outline_color}: A character string or vector of character strings
specifying the color of node outlines. If a vector, the colors will be
recycled in the order of the variables in the supplied data frame.
\item \code{node_fill_color}: A character string or vector of character strings
specifying the fill color of nodes. If a vector, the colors will be
recycled in the order of the variables in the supplied data frame.
\item \code{node_text_color}: A character string or vector of character strings
specifying the text color for node labels. If a vector, the colors will
be recycled in the order of the variables in the supplied data frame.
\item \code{node_text_size}: A numeric scalar specifying the text size for node
labels. Default value is 8.
\item \code{flow_text_color}: A character string or vector of character strings
specifying the text color for flow labels. If a vector, the colors will
be recycled in the order of the flows in the supplied data frame.
\item \code{flow_text_size}: A numeric scalar specifying the text size for flow
labels. Default value is 3.
\item \code{main_arrow_color}: A character string or vector of character strings
specifying the text color for non-interaction flow arrows.
If a vector, the colors will be recycled in the order of the flows
in the supplied data frame.
\item \code{main_arrow_linetype}: Either a numeric scalar or a character scalar
specifying the linetype for main arrows (non-interaction arrows). This
argument is passed to the \code{linetype} argument in ggplot2. From
the ggplot2 documentation: "The linetype aesthetic can be specified
with either an integer (0-6), a name (0 = blank, 1 = solid, 2 = dashed,
3 = dotted, 4 = dotdash, 5 = longdash, 6 = twodash), a mapping to a
discrete variable, or a string of an even number (up to eight) of
hexadecimal digits which give the lengths in consecutive positions in
the string." Default is 1 (solid).
\item \code{main_arrow_size}: A numeric scaler specifying the line size for the
main arrows (non-interaction arrows).
\item \code{interaction_arrow_color}: A character string or vector of character
strings specifying the text color for interaction flow arrows.
If a vector, the colors will be recycled in the order of the flows
in the supplied data frame.
\item \code{interaction_arrow_linetype}: Either a numeric scalar or a character scalar
specifying the linetype for interaction arrows. This
argument is passed to the \code{linetype} argument in ggplot2. From
the ggplot2 documentation: "The linetype aesthetic can be specified
with either an integer (0-6), a name (0 = blank, 1 = solid, 2 = dashed,
3 = dotted, 4 = dotdash, 5 = longdash, 6 = twodash), a mapping to a
discrete variable, or a string of an even number (up to eight) of
hexadecimal digits which give the lengths in consecutive positions in
the string." Default is 2 (dashed).
\item \code{interaction_arrow_size}: A numeric scalar specifying the line size for
the interaction arrows.
\item \code{with_grid} A logical indicating whether to return the ggplot
with a grid. Default is FALSE. The grid can be helpful if you
want/need to move items around.
}}
}
\value{
A ggplot2 object.
}
\description{
\code{make_diagram()} generates a \strong{ggplot2} object based on the data frames
made with \code{\link{prepare_diagram}}. The function only applies
aesthetics that are not associated with x, y locations. Colors, linetypes,
and other graphical options can be set by the user.
}
\examples{
varlabels <- c("S","I","R")
varnames <- c("Susceptible","Infected","Recovered")  # optional
flows <- list(S_flows = c("-b*S*I"),
              I_flows = c("b*S*I","-g*I"),
              R_flows = c("g*I"))
varlocations <-  matrix(data = c("S", "", "R", "", "I", "" ),
                        nrow = 2, ncol = 3, byrow = TRUE)
mymodel <- list(varlabels = varlabels, varnames = varnames,
flows = flows, varlocations = varlocations)
diagram_list <- prepare_diagram(model_list = mymodel)

# make diagram without grid
diagram <- make_diagram(diagram_list)

# make diagram with grid
diagram_with_grid <- make_diagram(diagram_list, diagram_settings = list(with_grid = TRUE))
}
