% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diagram.R
\name{make_diagram}
\alias{make_diagram}
\title{Make a ggplot2 model diagram.}
\usage{
make_diagram(df_list)
}
\arguments{
\item{df_list}{A list of data frames returned from the
\code{make_dataframes} function.}
}
\value{
A ggplot2 object.
}
\description{
Make a ggplot2 model diagram.
}
