% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diagram.R
\name{make_diagram}
\alias{make_diagram}
\title{Make a ggplot2 model diagram.}
\usage{
make_diagram(diagram_list, with_grid = FALSE)
}
\arguments{
\item{diagram_list}{A required list of data frames returned from the
\code{\link{prepare_diagram}} function and, optionally, updated with
\code{\link{update_diagram}}. See those functions for details
about this object.}

\item{with_grid}{A logical indicating whether to return the ggplot
with a grid. Default is FALSE. The grid can be helpful if you
want/need to move items around.}
}
\value{
A ggplot2 object.
}
\description{
\code{make_diagram()} generates a \strong{ggplot2} object based on the data frames
made with \code{\link{prepare_diagram}} and, optionally, updated with
\code{\link{update_diagram}}.
}
\details{
This function uses all the information in the data frames list
generated by \code{\link{prepare_diagram}} and, optionally, updated with
\code{\link{update_diagram}} to make a \code{ggplot2} object. All location
information and aesthetics are assumed fixed at this point -- no updates
are made within this function. The underlying \code{ggplot2} code can be
viewed by typing \code{make_diagram} with no parentheses in the R console.
}
\examples{
mymodel = list(variables = c("S","I","R"),
               flows = list(S_flows = c("-b*S*I"),
                            I_flows = c("b*S*I","-g*I"),
                            R_flows = c("g*I") ) )
diagram_list <- prepare_diagram(model_list = mymodel)

# make diagram without grid
diagram <- make_diagram(diagram_list)

# make diagram with grid
diagram_with_grid <- make_diagram(diagram_list, with_grid = TRUE)

}
