% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diagram.R
\name{make_diagram}
\alias{make_diagram}
\title{Make a ggplot2 model diagram.}
\usage{
make_diagram(df_list, with_grid = FALSE)
}
\arguments{
\item{df_list}{A list of data frames returned from the
\code{make_dataframes} function.}

\item{with_grid}{A logical indicating whether to return the ggplot
with a grid. Default is FALSE. The grid can be helpful if you
want/need to move items around.}
}
\value{
A ggplot2 object.
}
\description{
Make a ggplot2 model diagram.
}
