% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diagram.R
\name{make_diagram}
\alias{make_diagram}
\title{Make a ggplot2 model diagram.}
\usage{
make_diagram(
  diagram_list,
  diagram_settings = list(label_flows = TRUE, external_flows = TRUE, interaction_label
    = TRUE, var_outline_color = NA, var_fill_color = "#6aa4c8", var_text_color = "white",
    var_text_size = NA, flow_text_color = "black", flow_text_size = NA, main_flow_color =
    "grey25", main_flow_linetype = "solid", main_flow_size = 0.7, interaction_flow_color
    = "grey25", interaction_flow_linetype = "dashed", interaction_flow_size = 0.7,
    with_grid = FALSE)
)
}
\arguments{
\item{diagram_list}{A required list of data frames returned from the
\code{prepare_diagram} function. See that function for details
about this object.}

\item{diagram_settings}{An optional list of diagram aesthetic settings. The
following elements are supported and default values are provided:
\itemize{
\item \code{var_outline_color}: A character string or vector of character strings
specifying the color of variable outlines. If a vector, the colors will be
recycled in the order of the variables in the supplied data frame.
\item \code{var_fill_color}: A character string or vector of character strings
specifying the fill color of variables. If a vector, the colors will be
recycled in the order of the variables in the supplied data frame.
\item \code{var_text_color}: A character string or vector of character strings
specifying the text color for variable labels. If a vector, the colors will
be recycled in the order of the variables in the supplied data frame.
\item \code{var_text_size}: A numeric scalar specifying the text size for variable
labels. Default value is NA and the values in the \code{variables} data frame
returned by \code{\link{prepare_diagram}} are used. If a non-NA value is
supplied here, the values in the \code{variables} data frame are not used and
are replaced by the value specified here.

\item \code{main_flow_on}: A logical indicating if the main flow arrows should be plotted.
\item \code{main_flow_color}: A character string or vector of character strings
specifying the text color for non-interaction flow arrows.
If a vector, the colors will be recycled in the order of the flows
in the supplied data frame.
\item \code{main_flow_linetype}: Either a numeric scalar or a character scalar
specifying the linetype for main flows (non-interaction flows). This
argument is passed to the \code{linetype} argument in ggplot2. From
the ggplot2 documentation: "The linetype aesthetic can be specified
with either an integer (0-6), a name (0 = blank, 1 = solid, 2 = dashed,
3 = dotted, 4 = dotdash, 5 = longdash, 6 = twodash), a mapping to a
discrete variable, or a string of an even number (up to eight) of
hexadecimal digits which give the lengths in consecutive positions in
the string." Default is 1 (solid).
\item \code{main_flow_size}: A numeric scalar specifying the line size for the
main flows (non-interaction flows).
\item \code{main_flow_label_on}: A logical indicating if the labels for the main flows should be plotted.
\item \code{main_flow_label_color}:
\item \code{main_flow_label_size}:
\item \code{external_flow_on}: A logical indicating if the main flow arrows should be plotted.
\item \code{interaction_flow_on}: A logical indicating if the main flow arrows should be plotted.
#' \item \code{label_flows}: A logical indicating whether to label the flows
(TRUE, default) or not (FALSE).
\item \code{external_flows}: A logical indicating whether to include flows into
and out of the system (external flows). Default is TRUE (include).
\item \code{interaction_label}: A logical indicating whether to make the diagram
with interaction terms (typically curved flows leading to the
mid point of another flow) or to simply label the main flow. See
vignettes for examples.#'
\item \code{flow_text_color}: A character string or vector of character strings
specifying the text color for flow labels. If a vector, the colors will
be recycled in the order of the flows in the supplied data frame.
\item \code{flow_text_size}: A numeric scalar specifying the text size for flow
labels. Default value is NA and the values in the \code{flows} data frame
returned by \code{\link{prepare_diagram}} are used. If a non-NA value is
supplied here, the values in the \code{flows} data frame are not used and
are replaced by the value specified here.
\item \code{interaction_flow_color}: A character string or vector of character
strings specifying the text color for interaction flow arrows.
If a vector, the colors will be recycled in the order of the flows
in the supplied data frame.
\item \code{interaction_flow_linetype}: Either a numeric scalar or a character scalar
specifying the linetype for interaction flows. This
argument is passed to the \code{linetype} argument in ggplot2. From
the ggplot2 documentation: "The linetype aesthetic can be specified
with either an integer (0-6), a name (0 = blank, 1 = solid, 2 = dashed,
3 = dotted, 4 = dotdash, 5 = longdash, 6 = twodash), a mapping to a
discrete variable, or a string of an even number (up to eight) of
hexadecimal digits which give the lengths in consecutive positions in
the string." Default is 2 (dashed).
\item \code{interaction_flow_size}: A numeric scalar specifying the line size for
the interaction flows.
\item \code{with_grid} A logical indicating whether to return the ggplot
with a grid. Default is FALSE. The grid can be helpful if you
want/need to move items around.
}}
}
\value{
A ggplot2 object.
}
\description{
\code{make_diagram()} generates a \strong{ggplot2} object based on the data frames
made with \code{\link{prepare_diagram}}. The function only applies
aesthetics that are not associated with x, y locations. Colors, linetypes,
and other graphical options can be set by the user.
}
\examples{
varlabels <- c("S","I","R")
varnames <- c("Susceptible","Infected","Recovered")  # optional
flows <- list(S_flows = c("-b*S*I"),
              I_flows = c("b*S*I","-g*I"),
              R_flows = c("g*I"))
varlocations <-  matrix(data = c("S", "", "R", "", "I", "" ),
                        nrow = 2, ncol = 3, byrow = TRUE)
mymodel <- list(varlabels = varlabels, varnames = varnames,
flows = flows, varlocations = varlocations)
diagram_list <- prepare_diagram(model_list = mymodel)

# make diagram without grid
diagram <- make_diagram(diagram_list)

# make diagram with grid
diagram_with_grid <- make_diagram(diagram_list, diagram_settings = list(with_grid = TRUE))
}
