% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diagram.R
\name{make_diagram}
\alias{make_diagram}
\title{Make a ggplot2 model diagram.}
\usage{
make_diagram(
  diagram_list,
  diagram_settings = list(var_outline_color = NA, var_fill_color = "#6aa4c8",
    var_label_text = NA, var_label_on = TRUE, var_label_color = "white", var_label_size =
    NA, main_flow_on = TRUE, main_flow_color = "grey25", main_flow_linetype = "solid",
    main_flow_size = 0.7, main_flow_label_on = TRUE, main_flow_label_color = "black",
    main_flow_label_size = 5, interaction_flow_on = TRUE, interaction_flow_color =
    "grey25", interaction_flow_linetype = "dashed", interaction_flow_size = 0.7,
    interaction_flow_label_on = TRUE, interaction_flow_label_color = "black",     
    interaction_flow_label_size = 5, external_flow_on = TRUE, external_flow_color =
    "grey25", external_flow_linetype = "solid", external_flow_size = 0.7,
    external_flow_label_on = TRUE, external_flow_label_color = "black",
    external_flow_label_size = 5, with_grid = FALSE)
)
}
\arguments{
\item{diagram_list}{A required list of data frames returned from the
\code{prepare_diagram} function. See that function for details
about this object.}

\item{diagram_settings}{An optional list of diagram aesthetic settings. The
following elements are supported and default values are provided:
\itemize{
\item \code{var_outline_color}: A character string or vector of character strings
specifying the color of variable outlines. If a vector, the colors will be
recycled in the order of the variables in the supplied data frame.
\item \code{var_fill_color}: A character string or vector of character strings
specifying the fill color of variables. If a vector, the colors will be
recycled in the order of the variables in the supplied data frame.
\item \code{var_label_on}: A logical indicating if the labels for the variables
should be plotted.
\item \code{var_label_color}: A character string or vector of character strings
specifying the text color for variable labels. If a vector, the colors will
be recycled in the order of the variables in the supplied data frame.
\item \code{var_label_text}: A character vector the same length as the number
of variables in the \code{variables} data frame. If provided, these values
update the variable labels provided to \code{\link{prepare_diagram}}.
\item \code{var_label_size}: A numeric scalar specifying the text size for variable
labels. Note that any value supplied here overwrites
entries in the list structure returned by \code{\link{prepare_diagram}}.
Specifically, if you set this parameter when calling \code{\link{prepare_diagram}}
with \code{use_varnames = TRUE}, the value is used to compute box size,
but then the actual size of the label as provided here is applied.

\item \code{main_flow_on}: A logical indicating if the main flow arrows should be plotted.
\item \code{main_flow_color}: A character string or vector of character strings
specifying the text color for non-interaction flow arrows.
If a vector, the values will be recycled in the order of the flows
in the supplied data frame.
\item \code{main_flow_linetype}: Either a numeric scalar/vector or a character scalar/vector
specifying the linetype for main flows (non-interaction flows). This
argument is passed to the \code{linetype} argument in ggplot2. From
the ggplot2 documentation: "The linetype aesthetic can be specified
with either an integer (0-6), a name (0 = blank, 1 = solid, 2 = dashed,
3 = dotted, 4 = dotdash, 5 = longdash, 6 = twodash), a mapping to a
discrete variable, or a string of an even number (up to eight) of
hexadecimal digits which give the lengths in consecutive positions in
the string." Default is 1 (solid). If a vector, the values will
be recycled in the order of the flows in the supplied data frame.
\item \code{main_flow_size}: A numeric scalar or vector specifying the line size for the
main flows (non-interaction flows). If a vector, the values will
be recycled in the order of the flows in the supplied data frame.
\item \code{main_flow_label_on}: A logical indicating if the labels for the main
flows should be plotted.
\item \code{main_flow_label_color}: A character string or vector of character strings
specifying the text color for main flow labels. If a vector, the values will
be recycled in the order of the flows in the supplied data frame.
\item \code{main_flow_label_size}: A scalar or numeric vector
specifying the text size for main flow labels. If a vector, the values will
be recycled in the order of the flows in the supplied data frame.

\item \code{interaction_flow_on}: A logical indicating if the interaction flow arrows should be plotted.
\item \code{interaction_flow_color}: A character string or vector of character strings
specifying the text color for non-interaction flow arrows.
If a vector, the values will be recycled in the order of the flows
in the supplied data frame.
\item \code{interaction_flow_linetype}: Either a numeric scalar/vector or a character scalar/vector
specifying the linetype for interaction flows. This
argument is passed to the \code{linetype} argument in ggplot2. From
the ggplot2 documentation: "The linetype aesthetic can be specified
with either an integer (0-6), a name (0 = blank, 1 = solid, 2 = dashed,
3 = dotted, 4 = dotdash, 5 = longdash, 6 = twodash), a mapping to a
discrete variable, or a string of an even number (up to eight) of
hexadecimal digits which give the lengths in consecutive positions in
the string." Default is 1 (solid). If a vector, the values will
be recycled in the order of the flows in the supplied data frame.
\item \code{interaction_flow_size}: A numeric scalar or vector specifying the line size for the
interaction flows (non-interaction flows). If a vector, the values will
be recycled in the order of the flows in the supplied data frame.
\item \code{interaction_flow_label_on}: A logical indicating if the labels for the interaction
flows should be plotted.
\item \code{interaction_flow_label_color}: A character string or vector of character strings
specifying the text color for interaction flow labels. If a vector, the values will
be recycled in the order of the flows in the supplied data frame.
\item \code{interaction_flow_label_size}: A scalar or numeric vector
specifying the text size for interaction flow labels. If a vector, the values will
be recycled in the order of the flows in the supplied data frame.

\item \code{external_flow_on}: A logical indicating if the external flow arrows should be plotted.
\item \code{external_flow_color}: A character string or vector of character strings
specifying the text color for non-interaction flow arrows.
If a vector, the values will be recycled in the order of the flows
in the supplied data frame.
\item \code{external_flow_linetype}: Either a numeric scalar/vector or a character scalar/vector
specifying the linetype for external flows. This
argument is passed to the \code{linetype} argument in ggplot2. From
the ggplot2 documentation: "The linetype aesthetic can be specified
with either an integer (0-6), a name (0 = blank, 1 = solid, 2 = dashed,
3 = dotted, 4 = dotdash, 5 = longdash, 6 = twodash), a mapping to a
discrete variable, or a string of an even number (up to eight) of
hexadecimal digits which give the lengths in consecutive positions in
the string." Default is 1 (solid). If a vector, the values will
be recycled in the order of the flows in the supplied data frame.
\item \code{external_flow_size}: A numeric scalar or vector specifying the line size for the
external flows (non-interaction flows). If a vector, the values will
be recycled in the order of the flows in the supplied data frame.
\item \code{external_flow_label_on}: A logical indicating if the labels for the external
flows should be plotted.
\item \code{external_flow_label_color}: A character string or vector of character strings
specifying the text color for external flow labels. If a vector, the values will
be recycled in the order of the flows in the supplied data frame.
\item \code{external_flow_label_size}: A scalar or numeric vector
specifying the text size for external flow labels. If a vector, the values will
be recycled in the order of the flows in the supplied data frame.
\item \code{with_grid} A logical indicating whether to return the ggplot
with a grid. Default is FALSE. The grid can be helpful if you
want/need to move items around.
}}
}
\value{
A ggplot2 object.
}
\description{
\code{make_diagram()} generates a \strong{ggplot2} object based on the data frames
made with \code{\link{prepare_diagram}}. The function only applies
aesthetics that are not associated with x, y locations. Colors, linetypes,
and other graphical options can be set by the user.
}
\examples{
mymodel = list(varlabels = c("S","I","R"),
               flows = list(S_flows = c("-b*S*I"),
                            I_flows = c("b*S*I","-g*I"),
                            R_flows = c("g*I") ) )
diagram_list <- prepare_diagram(model_list = mymodel)

# make diagram without grid
diagram <- make_diagram(diagram_list)

# make diagram with grid
diagram_with_grid <- make_diagram(diagram_list, diagram_settings = list(with_grid = TRUE))
}
