% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_settings_df.R
\name{make_new_settings_df}
\alias{make_new_settings_df}
\title{Updates specified columns in a data frame with new values from a list.}
\usage{
make_new_settings_df(n, diagram_settings, list_names, column_names)
}
\arguments{
\item{n}{The number of variables or flows.}

\item{diagram_settings}{The diagram_settings list.}

\item{list_names}{The names of the list elements relevant for the current operation.}

\item{column_names}{The column names that map to each list name in diagram_settings.
Must be in the correct order.}
}
\description{
Helper function that updates all settings columns in a
flows or variables data frame based on the new values in diagram_settings.
}
