% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_diagram.R
\name{prepare_diagram}
\alias{prepare_diagram}
\title{Create data frames for plotting from model elements.}
\usage{
prepare_diagram(input_list)
}
\arguments{
\item{input_list}{A list of model elements. The list must contain at least
two elements with names \code{varlabels} and \code{flows}.}
}
\value{
A list of data frames.
}
\description{
Create data frames for plotting from model elements.
}
