% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_feedback_curvature.R
\name{set_feedback_curvature}
\alias{set_feedback_curvature}
\title{Sets the curvature values for feedback arrows}
\usage{
set_feedback_curvature(edf)
}
\arguments{
\item{edf}{Edges df}
}
\value{
A data frame
}
\description{
Sets the curvature values for feedback arrows
}
