% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tofroms.R
\name{update_tofroms}
\alias{update_tofroms}
\title{Updates the numeric to/from ids to the variable label}
\usage{
update_tofroms(targ, nodes)
}
\arguments{
\item{targ}{The target data frame, typically the flows df.}

\item{nodes}{The nodes data frame.}
}
\description{
Updates the numeric to/from ids to the variable label
}
