% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_diagram.R
\name{write_diagram}
\alias{write_diagram}
\title{Generate a reproducible R script to make the diagram.}
\usage{
write_diagram(
  model_list = NULL,
  diagram_list = NULL,
  diagram_settings = NULL,
  use_varnames = FALSE,
  with_grid = FALSE,
  directory = NULL,
  filename = "diagram_code.R",
  always_overwrite = FALSE
)
}
\arguments{
\item{model_list}{A \strong{flowdiagramr} input list. See
\code{\link{prepare_diagram}}. Also see \code{Details} below.}

\item{diagram_list}{A \strong{flowdiagramr} input structure, resulting from
a call to \code{\link{prepare_diagram}}. See \code{Details} below.}

\item{diagram_settings}{A named list of diagram aesthetics. See
\code{\link{make_diagram}} documentation. Default is \code{NULL} and the
default values from \code{\link{make_diagram}} are used.}

\item{use_varnames}{A logical indicating whether to label nodes with
variable abbreviations (\code{FALSE}; default) or to use the full names
provided in the \code{varnames} element of \code{model_list} (\code{TRUE}).}

\item{with_grid}{A logical indicating whether to return the ggplot
with a grid. Default is FALSE. The grid can be helpful if you
want/need to move items around.}

\item{directory}{File directory in which to save the produced R file.
Default location is the current working directory.}

\item{filename}{Name of the file, must end in '.R'. Default name is
'diagram_code.R'.}

\item{always_overwrite}{A logical indicating if you want to skip being asked
if you want to overwrite an already existing file.
Default is FALSE. Change to TRUE at own risk.}
}
\value{
R code written to a file as specified by settings.
Also, a message is returned telling the user where the file is.
}
\description{
\code{write_diagram()} generates code in the form of a stand-alone R script to
produce a diagram. By editing the generated code, the user can
make manual adjustments to the diagram.
}
\details{
You need to supply at least one of \code{model_list}
or \code{diagram_list}. If you supply both, \code{model_list} is included in the
resulting R file, but it is not used. Including it can be good just so
you have the complete model specification in one script.
}
\examples{
\dontrun{
varlabels <- c("S","I","R")
varnames <- c("Susceptible","Infected","Recovered")  # optional
flows <- list(S_flows = c("-b*S*I"),
              I_flows = c("b*S*I","-g*I"),
              R_flows = c("g*I"))
varlocations <-  matrix(data = c("S", "", "R", "", "I", "" ),
                        nrow = 2, ncol = 3, byrow = TRUE)
model_list <- list(varlabels = varlabels, varnames = varnames,
flows = flows, varlocations = varlocations)
diagram_list <- prepare_diagram(model_list = model_list)

# generate R code from model_list
write_diagram(model_list = model_list)

# generate R code from diagram_list
write_diagram(diagram_list = diagram_list)

#' # generate R code from both
write_diagram(model_list = model_list, diagram_list = diagram_list)
}
}
